services/user_service/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
PyJWT==2.8.0
python-decouple==3.8
django-environ==0.11.2
prometheus-client==0.19.0
structlog==23.2.0

